requirements.txt
Django==5.2.1
djangorestframework==3.16.0
django-cors-headers==4.7.0
gunicorn
psycopg2-binary==2.9.10
python-dotenv==0.19.2
pandas==2.2.3
openpyxl==3.1.5
whitenoise==6.5.0